requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sentence-transformers==2.3.1
faiss-cpu==1.7.4
pdfplumber==0.10.3
python-multipart==0.0.6
requests==2.31.0
pydantic-settings==2.1.0
pydantic==2.5.3
python-dotenv==1.0.0
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
ruff==0.1.14
black==24.1.1
mypy==1.8.0
